requirements.txt
Flask==1.0.2
Flask-Injector==0.12.0
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
requests
pyyaml
